flask_backend/requirements.txt
# PyTorch CPU-only build
--extra-index-url https://download.pytorch.org/whl/cpu

# Core ML dependencies (Synced for v2.6.0)
torch==2.6.0+cpu
torchvision==0.21.0+cpu
numpy==1.26.4

# Web Framework & Utils
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.4
python-dotenv==1.0.1
gunicorn==22.0.0

# Image Processing
Pillow==10.4.0
opencv-python-headless==4.10.0.84